+++ src/entities/name.ts
@@ -1,5 +1,5 @@
-import { Either, left, right } from '../shared/either'
-import { InvalidNameError } from './errors/invalid-name-error'
+import { Either, left, right } from '../shared'
+import { InvalidNameError } from './errors'
 
 export class Name {
   public readonly value: string
