+++ src/usecases/register-user-on-maling-list/ports/user-repository.ts
@@ -1,4 +1,4 @@
-import { UserData } from '../../../entities/user-data'
+import { UserData } from '../../../entities'
 
 export interface UserRepository {
   add(user: UserData): Promise<void>
