+++ src/usecases/register-user-on-maling-list/register-user-on-maling-list.ts
@@ -1,9 +1,7 @@
-import { InvalidEmailError } from '../../entities/errors/invalid-email-error'
-import { InvalidNameError } from '../../entities/errors/invalid-name-error'
-import { User } from '../../entities/user'
-import { UserData } from '../../entities/user-data'
-import { Either, left, right } from '../../shared/either'
-import { UserRepository } from './ports/user-repository'
+import { InvalidEmailError, InvalidNameError } from '../../entities/errors'
+import { User, UserData } from '../../entities'
+import { Either, left, right } from '../../shared'
+import { UserRepository } from './ports'
 
 export class RegisterUserOnMailingList {
     private readonly userRepo: UserRepository
